refactor: make integration tests their own workspace packages
