+++ pnpm-workspace.yaml
@@ -1,5 +1,4 @@
 packages:
   - 'packages/*'
   - 'examples/*'
-  - 'tests/*'
-  - 'tests/integration/packages/*'
+  - 'tests/integration/*'
