+++ tests/integration/durable-objects/tsconfig.integration.json
@@ -1,5 +1,7 @@
 {
 	"extends": "@vite-plugin-cloudflare/typescript-config/base.json",
-	"include": ["src"],
-	"exclude": ["src/fixtures"]
+	"include": ["*.test.ts"],
+	"compilerOptions": {
+		"noEmit": true
+	}
 }
