+++ tests/integration/durable-objects/worker-a/index.ts
