+++ tests/integration/durable-objects/worker-b/index.ts
