+++ tests/integration/durable-objects/worker-b/wrangler.toml
