+++ tests/integration/durable-objects/worker-c/index.ts
