+++ tests/integration/module-resolution/index.ts
