+++ tests/integration/module-resolution/packages/cloudflare-imports/index.d.ts
