+++ tests/integration/module-resolution/packages/cloudflare-imports/index.js
