+++ tests/integration/module-resolution/packages/cloudflare-imports/package.json
