+++ tests/integration/module-resolution/packages/requires/ext.d.ts
