+++ tests/integration/module-resolution/packages/requires/ext.js
