+++ tests/integration/module-resolution/packages/requires/hello.js
