+++ tests/integration/module-resolution/packages/requires/index.d.ts
