+++ tests/integration/module-resolution/packages/requires/index.js
