+++ tests/integration/module-resolution/packages/requires/json.d.ts
