+++ tests/integration/module-resolution/packages/requires/json.js
