+++ tests/integration/module-resolution/packages/requires/no-ext.d.ts
