+++ tests/integration/module-resolution/packages/requires/no-ext.js
