+++ tests/integration/module-resolution/packages/requires/package.json
