+++ tests/integration/module-resolution/packages/requires/world.cjs
