+++ tests/integration/module-resolution/src/cloudflare-imports.ts
