+++ tests/integration/module-resolution/src/external-cloudflare-imports.ts
@@ -1,4 +1,4 @@
-import { durableObjectName } from '@cloudflare-dev-module-resolution/cloudflare-imports';
+import { durableObjectName } from '../packages/cloudflare-imports';
 
 export default {
 	'(EXTERNAL) (cloudflare:workers) DurableObject.name': durableObjectName,
