+++ tests/integration/module-resolution/src/require-ext.ts
