+++ tests/integration/module-resolution/src/require-json.ts
