+++ tests/integration/module-resolution/src/require-no-ext.ts
