+++ tests/integration/module-resolution/tsconfig.fixtures.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@vite-plugin-cloudflare/typescript-config/worker.json",
+	"include": ["**/*.ts"],
+	"exclude": ["*.test.ts"],
+	"compilerOptions": {
+		"noEmit": true
+	}
+}
