+++ tests/integration/module-resolution/tsconfig.integration.json
@@ -1,0 +1,7 @@
+{
+	"extends": "@vite-plugin-cloudflare/typescript-config/base.json",
+	"include": ["*.test.ts"],
+	"compilerOptions": {
+		"noEmit": true
+	}
+}
