+++ tests/integration/service-bindings/worker-a/index.ts
