+++ tests/integration/service-bindings/worker-a/wrangler.toml
