+++ tests/integration/service-bindings/worker-b/index.ts
