+++ tests/integration/service-bindings/worker-b/wrangler.toml
