+++ tests/integration/test-helpers/package.json
@@ -1,15 +1,17 @@
 {
-	"name": "@examples/multi-worker",
+	"name": "@tests/integration-test-helpers",
 	"private": true,
 	"type": "module",
 	"scripts": {
-		"dev": "vite dev"
+		"types:check": "tsc"
 	},
 	"devDependencies": {
-		"@cloudflare/workers-types": "^4.20241004.0",
+		"@cloudflare/workers-types": "^4.20240925.0",
 		"@flarelabs-net/vite-plugin-cloudflare": "workspace:*",
 		"@vite-plugin-cloudflare/typescript-config": "workspace:*",
+		"typescript": "^5.6.2",
 		"vite": "6.0.0-beta.2",
-		"wrangler": "^3.80.4"
+		"vitest": "^2.1.2",
+		"wrangler": "^3.80.0"
 	}
 }
