+++ tests/integration/test-helpers/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "@vite-plugin-cloudflare/typescript-config/base.json",
+	"include": ["**/*.ts"],
+	"compilerOptions": {
+		"noEmit": true
+	}
+}
