--- tests/integration/tsconfig.fixtures.json
@@ -1,4 +1,0 @@
-{
-	"extends": "@vite-plugin-cloudflare/typescript-config/worker.json",
-	"include": ["src/fixtures"]
-}
